requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
torch==2.0.0
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.2.2
matplotlib==3.7.1
tqdm==4.65.0
Pillow==9.5.0
torchvision==0.15.1
pickle5==0.0.12